i prob have to rewrite my multimedia.dart

+++ app/lib/util/util.dart
@@ -97,7 +97,7 @@ Future<void> carefulShowDialog({required BuildContext context, required WidgetBu
   }
 }
 
-double getSongDurationValue(int position, int total) {
+double getSongDurationValue(double position, double total) {
   if (total == 0) return 0;
   var value = position / total;
   return value;
